canonicalize tableaus
The solver was wasting incredible amounts of time shuffling cards around
between equivalent states. Picture two long runs with an empty column: it
would try every permutation of runs on columns, then break up the runs to try
every permutation of cards in runs or the empty column. Then it would slightly
move some other card in another column and do it all again.

Sorting the cells and columns makes for an incredible reduction in the number
of states enumerated during a solving. Putting Dragon before Suited puts them
on the left because freeing them up for collection is a very positive move and
possibleMoves generates Pack commands from left-to-right.