120 moves in, timeout that tableau
If it can get to a Timeout state by another shorter path it'll keep solving.
This version doesn't actually need to track the states, it could just return
if the moveCount is too deep, but this leaves it open for future expansion to,
before finally returning a loss, solving without a timeout from the cancelled
states.